backend/pyproject.toml
[project]
name = "semita-finder-backend"
version = "0.1.0"
description = "API REST for semita finder app"
authors = [
    {name = "c2ksoria",email = "carlosvsoria1982@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-admin (>=2.0.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
